Battle config list screen ready for Navigation compose
